Refactored pkg name from external to clients
