+++ src/main/java/uk/gov/hmcts/reform/waworkflowapi/controllers/startworkflow/CreateTaskController.java
@@ -11,14 +11,14 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import uk.gov.hmcts.reform.waworkflowapi.clients.model.DmnValue;
+import uk.gov.hmcts.reform.waworkflowapi.clients.model.EvaluateDmnRequest;
+import uk.gov.hmcts.reform.waworkflowapi.clients.model.EvaluateDmnResponse;
+import uk.gov.hmcts.reform.waworkflowapi.clients.model.SendMessageRequest;
+import uk.gov.hmcts.reform.waworkflowapi.clients.service.EvaluateDmnService;
+import uk.gov.hmcts.reform.waworkflowapi.clients.service.SendMessageService;
 import uk.gov.hmcts.reform.waworkflowapi.common.TaskToCreate;
 import uk.gov.hmcts.reform.waworkflowapi.duedate.DueDateService;
-import uk.gov.hmcts.reform.waworkflowapi.external.model.DmnValue;
-import uk.gov.hmcts.reform.waworkflowapi.external.model.EvaluateDmnRequest;
-import uk.gov.hmcts.reform.waworkflowapi.external.model.EvaluateDmnResponse;
-import uk.gov.hmcts.reform.waworkflowapi.external.model.SendMessageRequest;
-import uk.gov.hmcts.reform.waworkflowapi.external.service.EvaluateDmnService;
-import uk.gov.hmcts.reform.waworkflowapi.external.service.SendMessageService;
 
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
