+++ src/test/java/uk/gov/hmcts/reform/waworkflowapi/clients/service/TaskClientServiceTest.java
@@ -1,10 +1,10 @@
-package uk.gov.hmcts.reform.waworkflowapi.external.service;
+package uk.gov.hmcts.reform.waworkflowapi.clients.service;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.waworkflowapi.external.model.DmnValue;
-import uk.gov.hmcts.reform.waworkflowapi.external.model.EvaluateDmnRequest;
+import uk.gov.hmcts.reform.waworkflowapi.clients.model.DmnValue;
+import uk.gov.hmcts.reform.waworkflowapi.clients.model.EvaluateDmnRequest;
 
 import java.util.ArrayList;
 import java.util.List;
